♻️ Moved Cells into seperate folders
