+++ libraries/core-react/src/components/Table/DataCell/index.ts
@@ -1,0 +1,1 @@
+export * from './DataCell'
