+++ libraries/core-react/src/components/Table/HeaderCell/index.ts
@@ -1,0 +1,1 @@
+export * from './HeaderCell'
